app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.0",
    "pyarrow>=19.0.0",
    "plotly>=5.24.0,<6.0",
    "requests>=2.32.5",
"taipy-gui>=4.0.0",
]
